moved Data Class, renamed city_placeholder
