--- app/src/main/java/DataClass.java
@@ -1,25 +1,0 @@
-import com.pitonneux.android.learningcircleproject.R;
-
-/**
- * Created by jonathan on 16-07-09.
- */
-public class DataClass {
-
-//    Data lives here
-    public static int[] cityKeys = {0,1,2,3,4,5,6,7,8,9};
-    public static String[] cityNames = {"London", "Paris", "Montreal", "Beijing", "Berlin",
-            "city6", "city7", "city8", "city9", "city10"};
-    public static String[] cityGeolocations = {"51°30′26″N 0°7′39″W", "48°51′24″N 2°21′03″E",
-            "45°30′N 73°34′W", "39°55′N 116°23′E", "52°31′N 13°23′E", "city6",
-            "city7", "city8", "city9", "city10"};
-    public static int[] cityIconIDs = {R.drawable.cityPlaceholder, R.drawable.cityPlaceholder,
-            R.drawable.cityPlaceholder, R.drawable.cityPlaceholder, R.drawable.cityPlaceholder,
-            R.drawable.cityPlaceholder, R.drawable.cityPlaceholder, R.drawable.cityPlaceholder,
-            R.drawable.cityPlaceholder, R.drawable.cityPlaceholder};
-    public static String[] descriptions = {"cool place", "cool place", "cool place", "cool place",
-            "cool place", "cool place", "cool place", "cool place", "cool place", "cool place"};
-    public static String[] countries = {"England", "France", "Canada", "China", "Germany",
-            "city6", "city7", "city8", "city9", "city10"};
-
-
-}
