+++ app/src/main/java/com/pitonneux/android/learningcircleproject/DataClass.java
@@ -1,0 +1,27 @@
+package com.pitonneux.android.learningcircleproject;
+
+/**
+ * Created by jonathan on 16-07-09.
+ */
+public class DataClass {
+
+
+//    Data lives here
+
+    public static int[] cityKeys = {0,1,2,3,4,5,6,7,8,9};
+    public static String[] cityNames = {"London", "Paris", "Montreal", "Beijing", "Berlin",
+            "city6", "city7", "city8", "city9", "city10"};
+    public static String[] cityGeolocations = {"51°30′26″N 0°7′39″W", "48°51′24″N 2°21′03″E",
+            "45°30′N 73°34′W", "39°55′N 116°23′E", "52°31′N 13°23′E", "city6Geo",
+            "city7Geo", "city8Geo", "city9Geo", "city10Geo"};
+    public static int[] cityIconIDs = {R.drawable.city_placeholder, R.drawable.city_placeholder,
+            R.drawable.city_placeholder, R.drawable.city_placeholder, R.drawable.city_placeholder,
+            R.drawable.city_placeholder, R.drawable.city_placeholder, R.drawable.city_placeholder,
+            R.drawable.city_placeholder, R.drawable.city_placeholder};
+    public static String[] descriptions = {"Lots of tea!", "Baguette!", "Poutine!", "Rice!",
+            "Beer!", "desc6", "desc7", "desc8", "desc9", "desc10"};
+    public static String[] countries = {"England", "France", "Canada", "China", "Germany",
+            "city6", "city7", "city8", "city9", "city10"};
+
+
+}
