+++ app/src/main/res/drawable/city_placeholder.png
